test(protocol): move sig permissioning tests to ZkAsset.js
delete ZkAssetPermissioning.js and move the tests to ZkAsset.js
done to follow test standard
