+++ packages/protocol/test/ERC1724/ZkAsset.js
@@ -1,6 +1,7 @@
 /* global artifacts, expect, contract, beforeEach, it:true */
 // ### External Dependencies
 const BN = require('bn.js');
+const crypto = require('crypto');
 
 // ### Internal Dependencies
 // eslint-disable-next-line object-curly-newline
@@ -9,7 +10,7 @@ const { constants, proofs } = require('@aztec/dev-utils');
 const secp256k1 = require('@aztec/secp256k1');
 const typedData = require('@aztec/typed-data');
 const truffleAssert = require('truffle-assertions');
-const { keccak256 } = require('web3-utils');
+const { keccak256, padLeft } = require('web3-utils');
 
 const { JOIN_SPLIT_PROOF } = proofs;
 const { note, proof, signer } = aztec;
@@ -30,31 +31,27 @@ const computeDomainHash = (validatorAddress) => {
     return keccak256(`0x${typedData.encodeMessageData(types, 'EIP712Domain', domain)}`);
 };
 
-contract('ZkAsset', (accounts) => {
+contract.only('ZkAsset', (accounts) => {
     let ace;
-    let aztecAccounts = [];
     let aztecJoinSplit;
     const canAdjustSupply = false;
     const canConvert = true;
     let erc20;
-    let notes = [];
     const scalingFactor = new BN(10);
     const tokensTransferred = new BN(100000);
-    let zkAsset;
 
     beforeEach(async () => {
         ace = await ACE.new({ from: accounts[0] });
-        aztecAccounts = [...new Array(10)].map(() => secp256k1.generateAccount());
-        notes = await Promise.all([
-            ...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, i * 10)),
-            ...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, i * 10)),
-        ]);
+        // aztecAccounts = [...new Array(10)].map(() => secp256k1.generateAccount());
+        // notes = await Promise.all([
+        //     ...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, i * 10)),
+        //     ...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, i * 10)),
+        // ]);
         await ace.setCommonReferenceString(constants.CRS);
         aztecJoinSplit = await JoinSplit.new();
         await ace.setProof(JOIN_SPLIT_PROOF, aztecJoinSplit.address);
 
         erc20 = await ERC20Mintable.new();
-        zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
 
         await Promise.all(
             accounts.map((account) => {
@@ -72,12 +69,14 @@ contract('ZkAsset', (accounts) => {
 
     describe('Success States', async () => {
         it('should compute the correct domain hash', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
             const domainHash = computeDomainHash(zkAsset.address);
             const result = await zkAsset.EIP712_DOMAIN_HASH();
             expect(result).to.equal(domainHash);
         });
 
         it('should set the flags', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
             const result = await zkAsset.flags();
             expect(result.active).to.equal(true);
             expect(result.canAdjustSupply).to.equal(false);
@@ -85,16 +84,23 @@ contract('ZkAsset', (accounts) => {
         });
 
         it('should set the linked token', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
             const result = await zkAsset.linkedToken();
             expect(result).to.equal(erc20.address);
         });
 
         it('should set the scaling factor', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
             const result = await zkAsset.scalingFactor();
             expect(result.toNumber()).to.equal(scalingFactor.toNumber());
         });
 
         it('should update a note registry with output notes', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(2)].map(() => secp256k1.generateAccount());
+            const noteValues = [0, 10];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
             const transferAmount = 10;
             const transferAmountBN = new BN(transferAmount);
 
@@ -124,6 +130,11 @@ contract('ZkAsset', (accounts) => {
         });
 
         it('should update a note registry by consuming input notes, with kPublic negative', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(4)].map(() => secp256k1.generateAccount());
+            const noteValues = [0, 10, 20, 30];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
             const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
                 inputNotes: [],
                 outputNotes: notes.slice(0, 2),
@@ -164,9 +175,14 @@ contract('ZkAsset', (accounts) => {
         });
 
         it('should update a note registry by consuming input notes, with kPublic positive', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(4)].map(() => secp256k1.generateAccount());
+            const noteValues = [70, 60, 50, 40];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
             const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
                 inputNotes: [],
-                outputNotes: notes.slice(6, 8),
+                outputNotes: notes.slice(0, 2),
                 senderAddress: accounts[0],
                 inputNoteOwners: [],
                 publicOwner: accounts[2],
@@ -181,10 +197,10 @@ contract('ZkAsset', (accounts) => {
             const withdrawalAmountBN = new BN(withdrawalAmount);
 
             const withdrawalAndTransferProof = proof.joinSplit.encodeJoinSplitTransaction({
-                inputNotes: notes.slice(6, 8),
-                outputNotes: notes.slice(4, 6),
+                inputNotes: notes.slice(0, 2),
+                outputNotes: notes.slice(2, 4),
                 senderAddress: accounts[0],
-                inputNoteOwners: aztecAccounts.slice(6, 8),
+                inputNoteOwners: aztecAccounts.slice(0, 2),
                 publicOwner: accounts[2],
                 kPublic: withdrawalAmount,
                 validatorAddress: zkAsset.address,
@@ -212,6 +228,11 @@ contract('ZkAsset', (accounts) => {
         });
 
         it('should update a note registry with kPublic = 0', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(4)].map(() => secp256k1.generateAccount());
+            const noteValues = [5, 17, 13, 25];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
             const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
                 inputNotes: [],
                 outputNotes: [notes[0], notes[3]],
@@ -245,6 +266,11 @@ contract('ZkAsset', (accounts) => {
 
     describe('Failure States', async () => {
         it('should fail if the ace fails to validate the proof', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(2)].map(() => secp256k1.generateAccount());
+            const noteValues = [0, 10];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
             const transferAmount = 10;
 
             const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
@@ -264,6 +290,232 @@ contract('ZkAsset', (accounts) => {
             const malformedProofData = `0x0123${depositProof.proofData.slice(6)}`;
             // no error message because it throws in assembly
             await truffleAssert.reverts(zkAsset.confidentialTransfer(malformedProofData, depositProof.signatures));
+        });
+
+        it('should fail if signatures are zero', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(4)].map(() => secp256k1.generateAccount());
+            const noteValues = [10, 20, 5, 25];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
+            const firstTransferAmount = 30;
+
+            const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [],
+                outputNotes: notes.slice(0, 2),
+                senderAddress: accounts[0],
+                inputNoteOwners: [],
+                publicOwner: accounts[0],
+                kPublic: firstTransferAmount * -1,
+                validatorAddress: zkAsset.address,
+            });
+
+            const depositProofOutput = outputCoder.getProofOutput(depositProof.expectedOutput, 0);
+            const depositProofHash = outputCoder.hashProofOutput(depositProofOutput);
+
+            const secondTransferAmount = 0;
+            const noteTransfer = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: notes.slice(0, 2),
+                outputNotes: notes.slice(2, 4),
+                senderAddress: accounts[0],
+                inputNoteOwners: aztecAccounts.slice(0, 2),
+                publicOwner: accounts[1],
+                kPublic: secondTransferAmount,
+                validatorAddress: zkAsset.address,
+            });
+
+            const noteTransferProofOutput = outputCoder.getProofOutput(noteTransfer.expectedOutput, 0);
+            const noteTransferProofHash = outputCoder.hashProofOutput(noteTransferProofOutput);
+
+            await ace.publicApprove(zkAsset.address, depositProofHash, firstTransferAmount, { from: accounts[0] });
+            await ace.publicApprove(zkAsset.address, noteTransferProofHash, secondTransferAmount, { from: accounts[1] });
+
+            await zkAsset.confidentialTransfer(depositProof.proofData, depositProof.signatures);
+
+            const length = 64;
+            const zeroSignature = new Array(length).fill(0).join('');
+            const zeroSignatures = `0x${zeroSignature + zeroSignature + zeroSignature}`;
+
+            await truffleAssert.reverts(zkAsset.confidentialTransfer(noteTransfer.proofData, zeroSignatures));
+        });
+
+        it('should fail if fake signatures are provided', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(4)].map(() => secp256k1.generateAccount());
+            const noteValues = [10, 20, 5, 25];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
+            const firstTransferAmount = 30;
+
+            const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [],
+                outputNotes: notes.slice(0, 2),
+                senderAddress: accounts[0],
+                inputNoteOwners: [],
+                publicOwner: accounts[0],
+                kPublic: firstTransferAmount * -1,
+                validatorAddress: zkAsset.address,
+            });
+
+            const depositProofOutput = outputCoder.getProofOutput(depositProof.expectedOutput, 0);
+            const depositProofHash = outputCoder.hashProofOutput(depositProofOutput);
+
+            const secondTransferAmount = 0;
+            const noteTransfer = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: notes.slice(0, 2),
+                outputNotes: notes.slice(2, 4),
+                senderAddress: accounts[0],
+                inputNoteOwners: aztecAccounts.slice(0, 2),
+                publicOwner: accounts[1],
+                kPublic: secondTransferAmount,
+                validatorAddress: zkAsset.address,
+            });
+
+            const noteTransferProofOutput = outputCoder.getProofOutput(noteTransfer.expectedOutput, 0);
+            const noteTransferProofHash = outputCoder.hashProofOutput(noteTransferProofOutput);
+
+            await ace.publicApprove(zkAsset.address, depositProofHash, firstTransferAmount, { from: accounts[0] });
+            await ace.publicApprove(zkAsset.address, noteTransferProofHash, secondTransferAmount, { from: accounts[1] });
+
+            await zkAsset.confidentialTransfer(depositProof.proofData, depositProof.signatures);
+
+            const fakeSignature = padLeft(crypto.randomBytes(32).toString('hex'));
+            const fakeSignatures = `0x${fakeSignature + fakeSignature + fakeSignature}`;
+
+            await truffleAssert.reverts(zkAsset.confidentialTransfer(noteTransfer.proofData, fakeSignatures));
+        });
+
+        it('should fail if different note owner signs the transaction', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(4)].map(() => secp256k1.generateAccount());
+            const noteValues = [10, 20, 5, 25];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
+            const firstTransferAmount = 30;
+            const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [],
+                outputNotes: notes.slice(0, 2),
+                senderAddress: accounts[0],
+                inputNoteOwners: [],
+                publicOwner: accounts[0],
+                kPublic: firstTransferAmount * -1,
+                validatorAddress: zkAsset.address,
+            });
+
+            const depositProofOutput = outputCoder.getProofOutput(depositProof.expectedOutput, 0);
+            const depositProofHash = outputCoder.hashProofOutput(depositProofOutput);
+
+            const fakeInputNoteOwners = aztecAccounts.slice(2, 4);
+
+            const secondTransferAmount = 0;
+            const noteTransfer = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: notes.slice(0, 2),
+                outputNotes: notes.slice(2, 4),
+                senderAddress: accounts[0],
+                inputNoteOwners: fakeInputNoteOwners,
+                publicOwner: accounts[1],
+                kPublic: secondTransferAmount,
+                validatorAddress: zkAsset.address,
+            });
+
+            const noteTransferProofOutput = outputCoder.getProofOutput(noteTransfer.expectedOutput, 0);
+            const noteTransferProofHash = outputCoder.hashProofOutput(noteTransferProofOutput);
+
+            await ace.publicApprove(zkAsset.address, depositProofHash, firstTransferAmount, { from: accounts[0] });
+            await ace.publicApprove(zkAsset.address, noteTransferProofHash, secondTransferAmount, { from: accounts[1] });
+
+            await zkAsset.confidentialTransfer(depositProof.proofData, depositProof.signatures);
+
+            await truffleAssert.reverts(zkAsset.confidentialTransfer(noteTransfer.proofData, noteTransfer.signatures));
+        });
+
+        it('should fail if validator address is fake', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(4)].map(() => secp256k1.generateAccount());
+            const noteValues = [10, 20, 5, 25];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
+            const firstTransferAmount = 30;
+            const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [],
+                outputNotes: notes.slice(0, 2),
+                senderAddress: accounts[0],
+                inputNoteOwners: [],
+                publicOwner: accounts[0],
+                kPublic: firstTransferAmount * -1,
+                validatorAddress: zkAsset.address,
+            });
+
+            const depositProofOutput = outputCoder.getProofOutput(depositProof.expectedOutput, 0);
+            const depositProofHash = outputCoder.hashProofOutput(depositProofOutput);
+
+            const fakeInputNoteOwners = aztecAccounts.slice(2, 4);
+
+            const randomAddress = proof.proofUtils.randomAddress();
+            const secondTransferAmount = 0;
+            const noteTransfer = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: notes.slice(0, 2),
+                outputNotes: notes.slice(2, 4),
+                senderAddress: accounts[0],
+                inputNoteOwners: fakeInputNoteOwners,
+                publicOwner: accounts[1],
+                kPublic: secondTransferAmount,
+                validatorAddress: randomAddress,
+            });
+
+            const noteTransferProofOutput = outputCoder.getProofOutput(noteTransfer.expectedOutput, 0);
+            const noteTransferProofHash = outputCoder.hashProofOutput(noteTransferProofOutput);
+
+            await ace.publicApprove(zkAsset.address, depositProofHash, firstTransferAmount, { from: accounts[0] });
+            await ace.publicApprove(zkAsset.address, noteTransferProofHash, secondTransferAmount, { from: accounts[1] });
+
+            await zkAsset.confidentialTransfer(depositProof.proofData, depositProof.signatures);
+
+            await truffleAssert.reverts(zkAsset.confidentialTransfer(noteTransfer.proofData, noteTransfer.signatures));
+        });
+
+        it('should fail if validator address is the joinSplit address', async () => {
+            const zkAsset = await ZkAsset.new(ace.address, erc20.address, scalingFactor, canAdjustSupply, canConvert);
+            const aztecAccounts = [...new Array(4)].map(() => secp256k1.generateAccount());
+            const noteValues = [10, 20, 5, 25];
+            const notes = await Promise.all([...aztecAccounts.map(({ publicKey }, i) => note.create(publicKey, noteValues[i]))]);
+
+            const firstTransferAmount = 30;
+            const depositProof = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [],
+                outputNotes: notes.slice(0, 2),
+                senderAddress: accounts[0],
+                inputNoteOwners: [],
+                publicOwner: accounts[0],
+                kPublic: firstTransferAmount * -1,
+                validatorAddress: zkAsset.address,
+            });
+
+            const depositProofOutput = outputCoder.getProofOutput(depositProof.expectedOutput, 0);
+            const depositProofHash = outputCoder.hashProofOutput(depositProofOutput);
+
+            const fakeInputNoteOwners = aztecAccounts.slice(2, 4);
+
+            const secondTransferAmount = 0;
+            const noteTransfer = proof.joinSplit.encodeJoinSplitTransaction({
+                inputNotes: notes.slice(0, 2),
+                outputNotes: notes.slice(2, 4),
+                senderAddress: accounts[0],
+                inputNoteOwners: fakeInputNoteOwners,
+                publicOwner: accounts[1],
+                kPublic: secondTransferAmount,
+                validatorAddress: aztecJoinSplit.address,
+            });
+
+            const noteTransferProofOutput = outputCoder.getProofOutput(noteTransfer.expectedOutput, 0);
+            const noteTransferProofHash = outputCoder.hashProofOutput(noteTransferProofOutput);
+
+            await ace.publicApprove(zkAsset.address, depositProofHash, firstTransferAmount, { from: accounts[0] });
+            await ace.publicApprove(zkAsset.address, noteTransferProofHash, secondTransferAmount, { from: accounts[1] });
+
+            await zkAsset.confidentialTransfer(depositProof.proofData, depositProof.signatures);
+
+            await truffleAssert.reverts(zkAsset.confidentialTransfer(noteTransfer.proofData, noteTransfer.signatures));
         });
     });
 });
